Refactor GraphTypes to outside
