+++ Libplanet.Explorer/GraphTypes/PlainValueKeyValuePair.cs
@@ -1,0 +1,10 @@
+using System;
+
+namespace Libplanet.Explorer.GraphTypes
+{
+    public class PlainValueKeyValuePair : Object
+    {
+        public string Key { get; set; }
+        public object Value { get; set; }
+    }
+}
