+++ Libplanet.Explorer/GraphTypes/PlainValueKeyValuePairType.cs
@@ -1,0 +1,13 @@
+using GraphQL.Types;
+
+namespace Libplanet.Explorer.GraphTypes
+{
+    public class PlainValueKeyValuePairType : ObjectGraphType<PlainValueKeyValuePair>
+    {
+        public PlainValueKeyValuePairType()
+        {
+            Field<StringGraphType>("Key");
+            Field<StringGraphType>("Value");
+        }
+    }
+}
